% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the sound an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{a string}

\item{sound}{a sting}
}
\value{
a string with the animal and sound
}
\description{
This is a package that print \strong{animals} and \strong{sounds}
It is a wrapper to \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog","woof")
}
