% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Print Animal Sounds}
\description{
Provides simple tools for combining animal names and the sounds they make into short character strings. It also includes example data for practicing package development workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lorenzo-Vives/animalsounds}
  \item \url{https://lorenzo-vives.github.io/animalsounds/}
  \item Report bugs at \url{https://github.com/Lorenzo-Vives/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Lorenzo Vives \email{lorenzo_vives@uc.cl}

Authors:
\itemize{
  \item Lorenzo Vives \email{lorenzo_vives@uc.cl}
}

}
\keyword{internal}
