% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{check_arg}
\alias{check_arg}
\title{Prints the Sound an Animal Makes}
\usage{
check_arg(arg, n = 1)
}
\arguments{
\item{animal}{A character vector of length 1}

\item{sound}{A character vector of length 1}
}
\value{
A string that combines the animal name and sound
}
\description{
This package prints the sound an animal makes, by creating a wrapper around \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
