% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataname.R
\docType{data}
\name{dataname}
\alias{dataname}
\title{Letters of the Roman Alphabet with Indices}
\format{
A data frame with 26 rows and 2 variables:
\describe{
\item{letter}{The letter as a character string.}
\item{index}{The corresponding numeric index.}
}
}
\usage{
dataname
}
\description{
A dataset of lower-case letters of the Roman alphabet and their
numeric index from a = 1 to z = 26.
}
\keyword{datasets}
