% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exited_sound.R
\name{excited_sound}
\alias{excited_sound}
\title{Make an excited animal sound}
\usage{
excited_sound(animal, sound, times = 3)
}
\arguments{
\item{animal}{A character string with an animal.}

\item{sound}{A character string containing an animal sound.}

\item{times}{Number of times to repeat the sound.}
}
\value{
A character vector.
}
\description{
Repeat an animal sound in uppercase.
}
\examples{
excited_sound("cow", "moo", 3)

}
