% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{first_n}
\alias{first_n}
\title{Get First N Rows of Data Frame}
\usage{
first_n(data, n)
}
\arguments{
\item{data}{A data frame}

\item{n}{An integer specifying the number of rows to return}
}
\description{
Return the first n rows of a data frame
}
