% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_indices.R
\docType{data}
\name{letter_indices}
\alias{letter_indices}
\title{Letters of the Alphabet with Indices}
\format{
A data frame with 26 rows and 2 variables:
\describe{
\item{letter}{The letters of the alphabet.}
\item{index}{The corresponding numeric index.}
}
}
\usage{
letter_indices
}
\description{
A dataset of lower-case letters of the Roman alphabet and their
numeric index from a = 1 to z = 26.
}
\keyword{datasets}
