% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: Sounds That Animals Make}
\description{
Tells us what sounds animals make.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Warwick-Stats-Resources/mypackage}
  \item Report bugs at \url{https://github.com/Warwick-Stats-Resources/mypackage/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.m.kaye@warwick.ac.uk}

}
\keyword{internal}
