% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print a sentence saying what sound an animal makes from user inputs}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{input from the user}

\item{sound}{input from the user}
}
\value{
a string that prints "the \emph{animal} goes \emph{sound}, \emph{sound}!"
}
\description{
This function uses \code{\link[base:paste]{base::paste0()}} to create a sentence that describes the sound an animal makes. You provide the \strong{animal} and the \strong{sound} as inputs.
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("cat", "meow")
}
