% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print a sentence saying what sound an animal makes from user inputs}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{input from the user}

\item{sound}{input from the user}
}
\value{
a string that prints "the {animal} goes {sound}, {sound}!"
}
\description{
Print a sentence saying what sound an animal makes from user inputs
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("cat", "meow")
}
