% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the sound that an animal makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string, representing an animal}

\item{sound}{A string, representing an animal sound}
}
\value{
A string concatentating the given animal and sound
}
\description{
Provides a simple wrapper around \code{\link[=paste0]{paste0()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Please note that this is a \strong{toy package}!
}
\examples{
animal <- "lion"
sound <- "roar"
animal_sounds(animal, sound)
"The lion goes roar!"


}
