% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Match the animal to the sound}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string, representing an animal}

\item{sound}{A string, representing an animal sound}
}
\value{
A string concatentating the given animal and sound
}
\description{
Match the animal to the sound
}
\examples{
animal <- "lion"
sound <- "roar"
animal_sounds(animal, sound)
"The lion goes roar!"
}
