% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Title}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{Name of the animal being discussed}

\item{sound}{The sound made by the animal}
}
\value{
A constructed sentence that includes the name of the animal and the sound that it makes
}
\description{
Description here
}
\details{
Details here, possibly several paragraphs. You can also explicitly use the
tag inside the main block if you want to separate it, but generally,
any paragraph after the Description becomes the details.

Here's a trail sentence linking to the \link[pkg:tidyplots]{\strong{Tidyplots}} package
that I enjoy using as a "gateway \strong{ggplot2}"
}
\examples{
animal_sounds("dog", "woof")
}
