% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sounds that Animals Make}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string containing the animal name. Must be longer than 1 character.}

\item{sound}{A string containing the sound. Must be longer than 1 character.}
}
\value{
A string in the form "The \{animal\} says \{sound\}!"
}
\description{
Print the sound that an animal makes in the console.
}
\examples{
animal_sounds("dog", "woof")
}
