% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sounds that Animals Make}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A \strong{string} containing an animal name. Must be longer than 1 character.}

\item{sound}{A \strong{string} containing the sound that the animal makes. Must be longer than 1 character.}
}
\value{
A \strong{string} in the form "The \{animal\} says \{sound\}!"
}
\description{
Print the sound that an animal makes in the console.
}
\details{
Is a wrapper to \code{\link[=paste0]{paste0()}}. This is a \emph{silly} function in a \strong{toy package}.
}
\examples{
animal_sounds("dog", "woof")
}
