% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print an Animal Sound}
\usage{
animal_sounds(animal, country, sound)
}
\arguments{
\item{animal}{string; animal name}

\item{country}{string; the country the animal is from}

\item{sound}{sting; sound the animal makes}
}
\value{
A sentence string of what sound an animal makes
}
\description{
Print an Animal Sound
}
\examples{
animal_sounds("dog", "France", "LeWoof")


}
