% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Title}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{a character string}

\item{sound}{A caracterr string}
}
\description{
Title
}
\examples{
animal_sounds("dog", "woolf")
}
