% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Prints the Sounds an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character vector of length 1}

\item{sound}{A character vector of length 1}
}
\value{
A string
}
\description{
Prints the sound that an animal makes.
}
\examples{
animals_sounds("dog", "bark")

}
