% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{A Sentence That Describes The Sound An Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string}

\item{sound}{A character string}
}
\value{
A string sentence containing the parameter values.
}
\description{
A Sentence That Describes The Sound An Animal Makes
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("owl", "hoot")
}
