% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Prints what a given animal says}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{a string containing an animal name}

\item{sound}{a string containing the sound that the animal makes}
}
\value{
a printed versiib
}
\description{
Prints what the animal says \emph{(the sound it makes)} in the console, using
\code{\link[=paste0]{paste0()}}
}
\examples{
animal_sounds("dog", "woof")
}
