% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string}

\item{sound}{A string}
}
\value{
A string with the animal and sound
}
\description{
Takes an \strong{animal} and the \strong{sound} it makes and returns the combination.
It is a wrapper to \code{\link[=paste0]{paste0()}}
}
\examples{
animal_sounds("dog", "woof")
}
