% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print an animal's sounds and a random gentle insult}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string representing an animal}

\item{sound}{A string representing an animal sound}
}
\value{
A string of the form "The (animal) goes (sound)! You (gentle insult)."
}
\description{
Print an animal's sounds and a random gentle insult
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("chimp", "ugh ugh")
}
