% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print an animal's sounds and a random gentle insult}
\usage{
animal_sounds(animal, sound = NULL, seed = NA)
}
\arguments{
\item{animal}{A string representing an animal}

\item{sound}{A string representing an animal sound}

\item{seed}{A random seed that can be set for reproducibility}
}
\value{
A string of the form "The \{animal\} goes \{sound\}! You \{gentle insult\}."
}
\description{
Wraps \code{\link[=paste0]{paste0()}} and \code{\link[=sample]{sample()}}. Use \code{\link[=set.seed]{set.seed()}} for reproducible insults.
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("chimp", "hooo hooo wah-hu-wah-hu-wah-hu-wah-hu-wah-hu waaaaargh waaaaargh hu hu hu")
}
