% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Retrieve animal sounds}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string representing the animal}

\item{sound}{A character string representing the sound the animal makes}
}
\value{
A character string in the format "The \link{animal} says \link{sound}!"
}
\description{
This function takes an animal and a sound as inputs and returns a string
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("cat", "meow")
}
