% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string}

\item{sound}{A sting}
}
\value{
A string with the animal and sound
}
\description{
Prints the \strong{sound} that an \emph{animal} makes.
It is a wrapper to \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
