% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{What is the animal?}

\item{sound}{What sound does the animal make?}
}
\value{
A sentence that says "The animal goes sound!".
}
\description{
This is a function that gives back the animal and its
sound \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "aw-aw")
}
