% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Function that takes animal and sound as input parameters}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{character with name of an animal (only of length = 1)}

\item{sound}{character with sound of an animal (only of length = 1)}
}
\value{
character
}
\description{
The function uses \link{paste0} and returns string with description of what sounds make the animal
}
\examples{
animal_sounds("dog", "woof woof")
"The dog says woof woof!"
}
