% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{name of animal, takes in string value}

\item{sound}{sound made by the animal, takes in string value}
}
\description{
Animal Sounds
}
\examples{
animal_sounds ("pig", "oink")
}
