% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Create Animal Sounds Message}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{the name of the animal (character)}

\item{sound}{the sound of the animal (character)}
}
\description{
Create a message stating what sound an animal makes.
}
\details{
We have used \code{\link[=paste0]{paste0()}} but not \code{\link[glue:glue]{glue::glue()}}
}
\examples{
animal_sounds("cow", "moo")
}
