% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Create Animal Sounds Message}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{the name of the animal (character)}

\item{sound}{the sound of the animal (character)}
}
\description{
Create Animal Sounds Message
}
\examples{
animal_sounds("cow", "moo")
}
