% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Specifies an animal and its sound.}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{sound}{}
}
\value{
The name of animal and its sound.
}
\description{
Specifies an animal and its sound.
}
\examples{
animal_sounds("cat", "purrr")
}
