% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Specifies an animal and its sound.}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{Name of the animal}

\item{sound}{Sound of the animal}
}
\value{
The name of animal and its sound.
}
\description{
This function uses a \link{paste0} from \emph{base} R.
}
\examples{
animal_sounds("cat", "purrr")
}
