% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal sounds function
Insert animal and sound parameters into \strong{\code{\link[=animal_sounds]{animal_sounds()}}}, will output a sentence that contains these
parameters to tell you what sounds the animals make}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{a single string}

\item{sound}{a single string}
}
\value{
prints what the animal says
}
\description{
Animal sounds function
Insert animal and sound parameters into \strong{\code{\link[=animal_sounds]{animal_sounds()}}}, will output a sentence that contains these
parameters to tell you what sounds the animals make
}
\examples{
animal_sounds("dog", "bark")
}
