% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound that an Animal Makes}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{A character string giving the animal}

\item{sound}{A character string giving the sound}
}
\value{
A string giving the sound the animal makes
}
\description{
A wrapper around \code{\link[base:paste]{base::paste0()}}.
}
\examples{
animal_sounds("cow", "moo")
}
