% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Prints the sound an animal makes}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{A character vector of length 1 representing the animal.}

\item{sound}{A character vector of length 1 representing the sound the animal makes.}
}
\value{
A character string describing the sound an animal makes.
}
\description{
Prints the sound that an animal makes by creating a wrapper around \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
