% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Create a String Giving the Sound an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{a character string specifying an animal}

\item{sound}{a character string specifying the animal sound}
}
\value{
A string of the form "The animal says sound!"
}
\description{
Given an animal and a sound, create a string that expresses
this information using \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
