% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Title}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{a character string naming an animal}

\item{sound}{a character string giving the animal sound}
}
\value{
a string of the form "The animal says sound!"
}
\description{
Description here
}
\details{
This function uses \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
