% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{What sounds the animal makes}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{string}

\item{sound}{string}
}
\value{
A string stating what sound the animal makes
}
\description{
Prints out what sound the animal makes
}
\details{
Uses function \code{\link[=paste0]{paste0()}} no it is \strong{NOT} do anything \emph{useful}
}
\examples{
animal_sounds("World", "hello")
}
