% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Prints the Sound an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character vector of length 1}

\item{sound}{A character vector of length 1}
}
\value{
A string that combines animal name and sound.
}
\description{
Prints the \strong{sound} that an animal \emph{makes}, by creating a wrapper around \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
