% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print a Sentence Saying What Sound an Animal Makes}
\usage{
animal_sounds(animal, country, sound)
}
\arguments{
\item{animal}{A character string specifying the animal}

\item{country}{A character string specifying the country where the animal is found}

\item{sound}{A character string specifying an animal sound}
}
\value{
A character string that describes the sound the animal makes
}
\description{
Print a Sentence Saying What Sound an Animal Makes
}
\examples{

animal_sounds("dog", "France", "woof")

}
