% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print What the Animal Says
Print \code{\link[rlang:abort]{rlang::abort()}}}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{a single string}

\item{sound}{a single string}
}
\value{
Print
}
\description{
Print What the Animal Says
Print \code{\link[rlang:abort]{rlang::abort()}}
}
\examples{
animal_sounds("dogs","woof")
}
