% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sounds that Animals Make}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string - the name of an animal}

\item{sound}{A string - the sound the animal makes}
}
\value{
A string of the form "The \{animal\} says \{sound\}!"
}
\description{
Print the sound that an animal makes in the console
}
\examples{
animal_sounds("dog", "woof")
}
