% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{An R package to describe what an animal does}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{The name of the animal in strings}

\item{sound}{The sound produced by the animal in strings}
}
\value{
Print the full sentence
}
\description{
Prints what the animal does \emph{(the sound it makes)} in the console, \code{\link[=paste0]{paste0()}}.It doesn't use \code{\link[mvtnorm:Mvt]{mvtnorm::rmvt()}}
}
\examples{
animal_sounds("dog", "woof")
}
