% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound an Animal Makes}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{A character vector of length 1}

\item{sound}{A character vector of length 1}
}
\value{
A character that combines the animal name and sound.
}
\description{
This function takes an animal and the sound it makes,
and returns a formatted string by creating a wrapper around \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("dog", "woof")
}
