% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds Is Just Practising Package Building}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{type of animal}

\item{sound}{he soudn it makes, preferably the matching sound ;)}
}
\value{
A sentence with the animal and the osund pasted together
}
\description{
Animal Sounds Is Just Practising Package Building
}
\examples{
animal_sounds("cat", "miauw")
}
