% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds Is Just Practising Package Building}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{string with type of animal}

\item{sound}{string with the sound it makes, preferably the matching sound ;)}
}
\value{
A sentence with the animal and the sound pasted together
}
\description{
Animal Sounds Is Just Practising Package Building
}
\details{
This is a \strong{wrapper} around \code{\link[=paste0]{paste0()}}.
}
\examples{
animal_sounds("cat", "miauw")

}
