% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds Function}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string input e.g. "dog", "lion", etc.}

\item{sound}{Another string input e.g. "woof", "roar" etc.}
}
\value{
A string in the form of a sentence of what the animal says
}
\description{
Function takes two inputs (animal, sound), and produces a string of what the
animal sounds like
}
\examples{
animalsounds("dog", "woof")
"The dog says woof!"

}
