% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound than an Animal Makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string}

\item{sound}{A character string}
}
\value{
A string saying what sound the animal makes
}
\description{
Print the Sound than an Animal Makes
}
\examples{
animal_sounds("dog", "woof")
}
