% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound than an Animal Makes}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{A character string}

\item{sound}{A character string}
}
\value{
A string saying what sound the animal makes
}
\description{
Provides a simple wrapper around \code{\link[=paste0]{paste0()}}.
}
\details{
Please note that this is a \strong{toy package}!
}
\examples{
animal_sounds("dog", "woof")
}
