% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Tell us what sound an animal makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{the name of an animal (a single character string)}

\item{sound}{the sound made by an animal (a single character string)}
}
\value{
A string like this: "The \{animal\} goes \{sound\}!"
}
\description{
Print the sound that an animal makes in the console.
}
\details{
This is an alias to \code{\link[=paste0]{paste0()}} but not to \code{\link[DEoptimR:JDEoptim]{DEoptimR::JDEoptim()}} or \link[DEoptimR:JDEoptim]{DE}
}
\examples{
animal_sounds("peacock","BlEuRk")
}
