% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Tell us what sounds an animal makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{the name of an animal (a single character string)}

\item{sound}{the sound made by an animal (a single character string)}
}
\value{
A string like this: "The (animal) goes (sound)"
}
\description{
Tell us what sounds an animal makes
}
\examples{
animal_sounds("peacock","BlEuRk!")
}
