% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Put your title here}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string}

\item{sound}{A string}
}
\value{
A string with the animal and sound
}
\description{
Description
}
\examples{
animal_sounds("dog","woof")
}
