% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the Sound than an Animal makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string representing the animal}

\item{sound}{A character string representing the sound the animal makes}
}
\value{
A character string describing the sound the animal makes
}
\description{
have to \strong{insall} the package first \code{\link[base:paste]{base::paste0()}}
}
\examples{
animal_sounds("dog", "woof")
}
