% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Check Animal Sounds}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{The animal that you're trying to work the sound out for.}

\item{sound}{The sound you want to check against the animal.}
}
\value{
Returns a sentence to the effect of "The cow says moo!"
}
\description{
A function which allows you to vibe out whether various animals make a sound of your choice. This relies upon you having a vague awareness of what each animal sounds like. Makes use of the \code{\link[base:paste]{base::paste0()}} function to put \emph{everything} together (it's \strong{basically} a wrapper).
}
\examples{
animal_sounds("cow", "moo")
}
