% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Sounds that animals make}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string, the name of an animal}

\item{sound}{A string, the sound the animal makes}
}
\value{
A string, conveying the sound the animal makes.
}
\description{
Uses \code{\link[=paste0]{paste0()}} to create a sentence stating the sound that the animal makes.
}
\examples{
animal_sounds("dog", "woof")
}
