% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print What the Animal Says}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{a single string}

\item{sound}{a single string}
}
\value{
Prints what the animals says
}
\description{
Prints what the animal says \emph{(the sound it makes)} in the console, using
\code{\link[=paste0]{paste0()}}.
}
\examples{
animals_sounds("dog", "woof")
}
