% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Animal Sounds Sentence Maker}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{name for the animal}

\item{sound}{sound that the animal makes}
}
\value{
a sentence in the format of what the animal says
}
\description{
Outputs a sentence of what sounds a given animal makes.
}
\examples{
animal_sounds("dog", "woof")
}
