% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/3mmaRand/animalsounds}
  \item Report bugs at \url{https://github.com/3mmaRand/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Emma Rand \email{emma.rand@york.ac.uk} (\href{https://orcid.org/0000-0002-1358-8275}{ORCID})

}
\keyword{internal}
