% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Prints the Sounds that Animals Make}
\description{
TOY PACKAGE for RForwards 2025. (Simple wrapper around paste0 to generate some animal sounds!).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aforestsomewhere/animalsounds}
  \item \url{https://aforestsomewhere.github.io/animalsounds/}
  \item Report bugs at \url{https://github.com/aforestsomewhere/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: A. Kate O'Mahony \email{katie.fala@teagasc.ie} (\href{https://orcid.org/0000-0001-7589-0341}{ORCID})

}
\keyword{internal}
