% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Prints Animals and Sounds}
\description{
Prints the sounds that animals make. Also contains a data set of farm animals and their sounds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EllaKaye/animalsounds}
  \item Report bugs at \url{https://github.com/EllaKaye/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{hello@example.com} (\href{https://orcid.org/0000-0001-2345-6789}{ORCID})

}
\keyword{internal}
