% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Prints Sounds that Animals Make}
\description{
The animalsounds package validates user input and prints the sounds that animals make. The validation is done through internal functions and the printing is done using the animal_sounds() function which accepts animal and sound as input string. Ideal input is one animal and one sound.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hazelcyril/animalsounds}
  \item Report bugs at \url{https://github.com/hazelcyril/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Hazel Cyril \email{hazel.cyril@warwick.ac.uk}

}
\keyword{internal}
