% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Prints Animal Sounds in a Sentence}
\description{
A practice package that prints what sound an animal makes depending on the variables inputted by the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamieGeo/animalsounds}
  \item Report bugs at \url{https://github.com/JamieGeo/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Jamie Burgess \email{jamiegburgess@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
