% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Prints out what the animal says}
\description{
Checks you only have 1 animal and 1 sound and prints what the animal says. Also handles soundless animals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KimSummers/animalsounds}
  \item Report bugs at \url{https://github.com/KimSummers/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Kim Summers \email{kimberley.summers@warwick.ac.uk}

}
\keyword{internal}
