% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Animal Sounds}
\description{
Prints a note about what an animal does (sounds).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Matthewadeoye/animalsounds}
  \item Report bugs at \url{https://github.com/Matthewadeoye/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Matthew Adeoye \email{matthewadeoye9@gmail.com}

}
\keyword{internal}
