% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michaket/animalsounds}
  \item Report bugs at \url{https://github.com/michaket/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Michael Ketzer \email{michael.ketzer@unibas.ch} (\href{https://orcid.org/0000-0002-9633-3947}{ORCID})

}
\keyword{internal}
