% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Links Animals to their Sounds}
\description{
Connects animals with their sounds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Panos-Ts/animalsounds}
  \item \url{https://panos-ts.github.io/animalsounds/}
  \item Report bugs at \url{https://github.com/Panos-Ts/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Panos Tsapras \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
