% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimalSounds-package.R
\docType{package}
\name{AnimalSounds-package}
\alias{AnimalSounds}
\alias{AnimalSounds-package}
\title{AnimalSounds: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pweingaertner/AnimalSounds}
  \item Report bugs at \url{https://github.com/pweingaertner/AnimalSounds/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
