% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimalSounds-package.R
\docType{package}
\name{AnimalSounds-package}
\alias{AnimalSounds}
\alias{AnimalSounds-package}
\title{AnimalSounds: Prints Animals and Sounds}
\description{
Provides fantastic functionalities related to animals and sounds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pweingaertner/AnimalSounds}
  \item Report bugs at \url{https://github.com/pweingaertner/AnimalSounds/issues}
}

}
\author{
\strong{Maintainer}: Paul Weingärtner \email{hello@example.com}

}
\keyword{internal}
