% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds-package.R
\docType{package}
\name{animalsounds-package}
\alias{animalsounds}
\alias{animalsounds-package}
\title{animalsounds: Links Animals to their Sounds}
\description{
Connects animals with their sounds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Warwick-Stats-Resources/animalsounds}
  \item Report bugs at \url{https://github.com/Warwick-Stats-Resources/animalsounds/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.m.kaye@warwick.ac.uk}

}
\keyword{internal}
