% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataname.R
\docType{data}
\name{dataname}
\alias{dataname}
\title{Example Data}
\format{
A dataframe with 26 rows and two columns
\describe{
\item{\code{letter}}{a letter of the alphabet}
\item{\code{index}}{the corresponding index of each letter}
}
}
\usage{
dataname
}
\description{
A pointless dataset with letters and numbers
}
\keyword{datasets}
