% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm Animals and their Sounds}
\format{
a dataset with two variables and t5 rows:
\describe{
\item{animal}{The name of the animal}
\item{sound}{The sound the animal makes}
}
}
\usage{
farm_animals
}
\description{
A dataset of farm animals and the sounds they make.
}
\keyword{datasets}
