% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Sounds Made by Animals}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{animal}{The animal as a character string.}
\item{sound}{The sound made by the animal as a character string.}
}
}
\usage{
farm_animals
}
\description{
A dataset of 3 animals and their corresponding sounds
}
\keyword{datasets}
