% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Data for Animal Sounds}
\format{
A dataframe with 3 rows and two columns
\describe{
\item{\code{animal}}{an animal name}
\item{\code{sound}}{the corresponding sound for each animal}
}
}
\usage{
farm_animals
}
\description{
A test dataset with animals and the sound they make
}
\keyword{datasets}
