% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm Animal Sounds}
\format{
A data frame with 5 rows and 2 columns:
\describe{
\item{\code{animal}}{The type of farm animal.}
\item{\code{sound}}{The sound the farm animal makes.}
}
}
\source{
Generated in \code{data-raw/farm_animals.R}
}
\usage{
farm_animals
}
\description{
Farm animals and the sounds they make.
}
\keyword{datasets}
