% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm animal names, sounds and feed}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{name}{The animal name as a character string.}
\item{sound}{The animal sound as a character string.}
\item{feed}{The animal feed as a character string.}
}
}
\usage{
farm_animals
}
\description{
A dataset of farm animals
}
\keyword{datasets}
