% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm animals with sounds}
\format{
Data frame with 2 variables and 7 entries:
\describe{
\item{animal}{Animal as a string}
\item{sound}{Sound as a string}
}
}
\usage{
farm_animals
}
\description{
Creates a dataset for these
}
\keyword{datasets}
