% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm Animal Sounds}
\format{
A data frame with five rows and two columns
\describe{
\item{\code{animal}}{Type of farm animal}
\item{\code{sound}}{Sound made by the animal}
}
}
\usage{
farm_animals
}
\description{
Farm animals and the sounds they make.
}
\keyword{datasets}
