% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{The sound of various farm animals}
\format{
A dataframe with 3 rows
\describe{
\item{animal}{the name of the animal}
\item{sond}{a rough representation of the typical sound made by the animal}
}
}
\usage{
farm_animals
}
\description{
A dataframe with sound of animals
}
\keyword{datasets}
