% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm Animal Sounds}
\format{
A data frame with three rows and two columns
\describe{
\item{animal}{A string giving the animal name}
\item{sound}{A string for the animal sound}
}
}
\usage{
farm_animals
}
\description{
Farm animals and the sounds they make
}
\keyword{datasets}
