% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{A dataset of farm animals and the sounds they make.}
\format{
A brief data frame with 4 rows and 2 variables.
\describe{
\item{animals}{The animal as a character string.}
\item{sounds}{The corresponding sound as a character string.}
}
}
\usage{
farm_animals
}
\description{
A dataset of farm animals and the sounds they make.
}
\keyword{datasets}
