% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm animals and their sounds}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{animal}{An animal as a character string.}
\item{sound}{The corresponding sound.}
}
}
\usage{
farm_animals
}
\description{
A data frame of farmyard animals and the sounds they make.
}
\keyword{datasets}
