% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-letter_indices.R
\docType{data}
\name{letter_indices}
\alias{letter_indices}
\title{Letters of the Alphabet with their Indices}
\format{
A data frame with 26 rows and 2 variables:
\describe{
\item{letter}{The letter as a character string.}
\item{index}{The index as an integer.}
}
}
\usage{
letter_indices
}
\description{
A dataset of lower case letters with their numeric indices from a = 1 to z = 26
}
\keyword{datasets}
