% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letter_indices}
\alias{letter_indices}
\title{Letters of the Alphabet with indices}
\format{
A data frame with 26 rows and 2 variables:
\describe{
\item{letter}{The letters of the alphabet}
\item{index}{The corresponding numberic index}
}
}
\usage{
letter_indices
}
\description{
A dataset of lower-case letters
}
\keyword{datasets}
